+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* [UIRS-34](https://issues.folio.org/browse/UIRS-34) "Data synchronization schedule" field enhancements
+
 ## [2.0.0](https://github.com/folio-org/ui-remote-storage/tree/v2.0.0) (2021-06-17)
 [Full Changelog](https://github.com/folio-org/ui-remote-storage/compare/v1.0.0...v2.0.0)